requirements.txt
django==3.1.1
django-ckeditor==6.0.0